+++ packages/carbon-graphs/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Added
-  * Added preliminary logic for new prop `config.axis.x.allowCalibration` to update automatically axis limits to allow padding.
+  * Added preliminary logic for new prop `config.axis.x.allowCalibration` to automatically update axis limits to allow padding.
 
 ## 2.21.0 - (February 15, 2022)