+++ packages/carbon-graphs/CHANGELOG.md
@@ -3,9 +3,7 @@
 ## Unreleased
 
 * Added
-  * Added preliminary logic for new prop `config.axis.x.allowCalibration` to automatically update axis limits to allow padding.
-  * Finished implementation for `config.axis.x.allowCalibration` for a line graph with a numerical x-axis.
-  * Finished implementation for `config.axis.x.allowCalibration` for a line graph with a timeseries x-axis.
+  * Added allowCalibration support for the x-axis on a line graph.
 
 ## 2.21.0 - (February 15, 2022)