+++ packages/carbon-graphs/CHANGELOG.md
@@ -4,8 +4,7 @@
 
 * Added
   * Added allowCalibration support for the x-axis on a line graph.
-  * Finished implementation for `config.axis.x.allowCalibration` for a paired result graph with a numerical x-axis.
-  * Finished implementation for `config.axis.x.allowCalibration` for a paired result graph with a timeseries x-axis.
+  * Finished implementation for `config.axis.x.allowCalibration` for a paired result graph with a numerical and timeseries x-axis.
 
 ## 2.21.0 - (February 15, 2022)