+++ packages/carbon-graphs/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * Added
   * Added allowCalibration support for the x-axis on a line graph.
-  * Finished implementation for `config.axis.x.allowCalibration` for a paired result graph with a numerical x-axis.
+  * Finished implementation for `config.axis.x.allowCalibration` for a paired result graph with a numerical and timeseries x-axis.
   * Finished implementation for `config.axis.x.allowCalibration` for a scatter graph with a numerical and timeseries x-axis.
   * Finished implementation for `config.axis.x.allowCalibration` for a bar graph with a numerical and timeseries x-axis.