+++ packages/carbon-graphs/CHANGELOG.md
@@ -3,9 +3,9 @@
 ## Unreleased
 
 * Added
-  * Added allowCalibration support for the x-axis on a line graph.
-  * Finished implementation for `config.axis.x.allowCalibration` for a paired result graph with a numerical and timeseries x-axis.
-  * Finished implementation for `config.axis.x.allowCalibration` for a scatter graph with a numerical and timeseries x-axis.
+  * Added support for `config.axis.x.allowCalibration` for a line graph with a numerical and timeseries x-axis.
+  * Added support for `config.axis.x.allowCalibration` for a paired result graph with a numerical and timeseries x-axis.
+  * Added support for `config.axis.x.allowCalibration` for a scatter graph with a numerical and timeseries x-axis.
 
 ## 2.21.0 - (February 15, 2022)